Moved generic parts from M13T004 to MTM1M3Movements
